+++ tests/test_rest.py
@@ -607,8 +607,9 @@ def test_patch_zones_vol_delta(client):
     if z['id'] in range(6):
       assert z['vol_f'] - (zones[z['id']]['vol_f'] + 0.1) < 0.0001 and z["vol_f_overflow"] == 0
 
-  # test overflowing deltas
+  # update each zones volume by 100%
   rv = client.patch('/api/zones', json={'zones': [z['id'] for z in zones], 'update': {'vol_delta_f': -1.0}})
+
   assert rv.status_code == HTTPStatus.OK
   jrv = rv.json()
   assert len(jrv['zones']) >= 6
@@ -643,6 +644,7 @@ def test_patch_zones_vol_delta(client):
 
   # test precedence
   rv = client.patch('/api/zones', json={'zones': [z['id'] for z in zones], 'update': {'vol_delta_f': 1.0, "vol": amplipi.models.MIN_VOL_DB}})
+
   assert rv.status_code == HTTPStatus.OK
   jrv = rv.json()
   assert len(jrv['zones']) >= 6